Extract history commands module and use widget labels in history text
